Handle errors returned by store effectively
This fixes the bug where in the GetNode and GetPod handlers couldn't
differentiate between a non-existent node or pod and an error due to the
store not being able to retrieve the node or pod.

This is handled by utilising the custom error types defined in the node
and pod registries.