Allow user to provide etcd ports
Added ability to specify etcd peer and client ports when the api is
started. The client port is defaulted to 2379 and the peer port is
randomly choosen from the available free ports on the host.

The controller by default uses the port 2379 to connect to the etcd and
watch for changes.